Support deep merge of query mocks
Our current implementation of query mocking only allows the latest
version of a particular mock to be in effect. This means that each time
you mock an object, you need to provide all properties for that object
in its entirety. This creates a lot of overhead when we only want to
mock a specific property on a mock for testing purposes.

This change adds support for allowing subsequent mocks to override
properties on any mocking functions that existed prior to that mock.

For example, if you have the following base mock object with defaults
defined as:
```javascript
Person: {
  id: 1,
  firstName: "John",
  lastName: "Doe",
  age: 35
}
```
And you are running a test where you are testing logic that consumes
age, you may want to mock person without regard for what their name is:
```javascript
Person: {
  age: 21
}
```

In today's world, you would get back the following when consuming the
Person object:
```javascript
Person: {
  age: 21
}
```
With this change in place, you would instead get:
```javascript
Person: {
  id: 1,
  firstName: "John",
  lastName: "Doe",
  age: 21
}
```